+++ manifest.json
@@ -4,7 +4,7 @@
   "name": "Mastodon Friend Checker",
   "description": "This extension shows 'Following' if you've already followed a remote Mastodon web profile.",
   "version": "0.2.0",
-  "browser_action": {
+  "action": {
     "default_icon": "icon.png",
     "default_popup": "popup.html"
   },
@@ -16,7 +16,8 @@
     "https://*/*"
   ],
   "background": {
-    "service_worker": "background.js"
+    "service_worker": "background.js",
+    "type": "module"
   },
   "icons": {
     "16": "icon16.png",
